backend/requirements.txt
Django>=4.2
djangorestframework
WeasyPrint
python-docx
docxtpl
django-cors-headers